% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpc.R
\name{cpc_weekly_sst}
\alias{cpc_weekly_sst}
\title{Fetches weekly Nino SST series from NOAA-CPC.}
\usage{
cpc_weekly_sst(output_type = c("data.frame", "ts", "xts"))
}
\arguments{
\item{output_type}{The output type of the resulting series. Can be \code{"data.frame"}, \code{"ts"} or \code{"xts"}.}
}
\value{
Time series of index observations.
}
\description{
Fetches weekly Nino SST series from NOAA-CPC.
}
\examples{
cpc_weekly_sst()

cpc_weekly_sst("xts")

}
