% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psl.R
\name{psl_index_names}
\alias{psl_index_names}
\title{Returns a vector with available index names to retrieve using \code{"psl_monthly"}}
\usage{
psl_index_names()
}
\value{
A character vector
}
\description{
Returns a vector with available index names to retrieve using \code{"psl_monthly"}
}
\examples{
psl_index_names()

}
