% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psl.R
\name{psl_monthly}
\alias{psl_monthly}
\title{Fetches monthly NOAA-PSL series.}
\usage{
psl_monthly(index_names, output_type = c("data.frame", "ts", "xts"))
}
\arguments{
\item{index_names}{A character vector of index names, see \code{"psl_index_names"},}

\item{output_type}{The output type of the resulting series. Can be \code{"data.frame"}, \code{"ts"} or \code{"xts"}.}
}
\value{
Time series of index observations.
}
\description{
Fetches monthly NOAA-PSL series.
}
\examples{
psl_monthly("nina1.anom")

psl_monthly(c("soi", "nao"), output_type="xts")

}
